requirements.txt
couchbase>=4.2.0
openai>=1.12.0
streamlit>=1.31.0
python-dotenv>=1.0.0
requests>=2.31.0
plotly>=5.18.0
streamlit-shadcn-ui>=0.1.19
